메인현황판 - API 개발 #57 -> 메인현황판 시간별 검침율 적시율 수정 -> 기타 메인현황판 수정 -> 검침현황판 API 기능 추가 및 수정 -> 장애현황판 DCU 시간별 장애정보 기능 변경 -> 기타 현황판 관련 소스 수정

+++ Ami/src/main/java/com/cnu/ami/dashboard/models/UseDayHourAllListVO.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 public class UseDayHourAllListVO {
 
 	private int time;
-	private double use;
+	private float use;
 	private long lp;
 
 }
